Remove myinfo page helpers and move functionality into my info page
